+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 -----
 
 - Add a pull to refresh to the podcast list and profile view (#239)
+- Testing
 
 7.22
 -----